// src/api/generated/schemas/providerType.ts
/**
 * Generated by orval v7.12.2 🍺
 * Do not edit manually.
 * Notifiable API
 * A simple example ASP.NET Core Web API
 * OpenAPI spec version: v1
 */

export type ProviderType = (typeof ProviderType)[keyof typeof ProviderType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ProviderType = {
  Mailjet: 'mailjet',
  SendGrid: 'sendGrid',
  FCM: 'fcm',
  APNs: 'apNs',
  Twilio: 'twilio',
  OneSignal: 'oneSignal',
} as const;